Properly mark history items added without user gesture https://bugs.webkit.org/show_bug.cgi?id=247909 rdar://100579077
Reviewed by Geoffrey Garen.

We recently introduced logic to mark history items as "created by JavaScript
without a user gesture" so that those could be skipped when the user navigates
back/forward via the browser UI.

However, which history items we marked with such flag was a little unexpected
and didn't quite match what Chrome did to address the same problem.

In particular, if on URL1 and then calling `pushState(null, null, 'a')` in JS
without a user gesture. We would previously mark the HistoryItem for URL1 as
created without a user gesture, instead of marking the one for URL1#a. URL1#a
is the actual HistoryItem created by JS so it makes more sense to mark this
one (which also matches Chrome's behavior).

* Source/WebCore/loader/FrameLoader.cpp:
(WebCore::FrameLoader::loadInSameDocument):
* Source/WebCore/loader/HistoryController.cpp:
(WebCore::FrameLoader::HistoryController::pushState):
* Tools/TestWebKitAPI/Tests/WebKit/WKBackForwardList.mm:
(TEST):

Canonical link: https://commits.webkit.org/252432.826@safari-7614-branch