REGRESSION (Ventura): [ macOS ] 2X TestWebKitAPI.SOAuthorizationRedirect.DismissUIDuringMiniaturization (API-Tests) are constant failures https://bugs.webkit.org/show_bug.cgi?id=247924 rdar://problem/102343029
Reviewed by Darin Adler.

The test assumed that actions to NSWindows or NSApp reflect synchronously
back to the notifications. This is not the case.

Ensure that the test expects asynchronous operation.

Fix incorrect adoptNS() for notification center observers. The add function is
not a NS creation function, and as such it doesn't return an adoptable ref.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/SOAuthorizationTests.mm:
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/256727@main