Add numeric identifier to generated getter name https://bugs.webkit.org/show_bug.cgi?id=247424 rdar://101940115
Reviewed by Justin Michaud.

Adds a call to set the ECMA property name for a generated getter/setter
when the property name is a numeric identifier. With this change, we
include the property name in the name of the new getter/setter, i.e.
'get 0' instead of 'get '.

* JSTests/stress/numeric-getter-has-correct-name.js: Added.
(x.prototype.get 0):
(x):
* LayoutTests/js/script-tests/function-toString-vs-name.js:
* Source/JavaScriptCore/parser/ASTBuilder.h:
(JSC::ASTBuilder::createGetterOrSetterProperty):

Canonical link: https://commits.webkit.org/256740@main