Fix debug assertion failure in IndentOutdentCommand::indentIntoBlockquote https://bugs.webkit.org/show_bug.cgi?id=240377 rdar://93236442
Reviewed by Ryosuke Niwa.

When indenting an element into a new uneditable blockquote we end up
replacing the element with an empty blockquote instead of indenting.
This also fails an assert in debug builds.
We should just bail out instead of trying to indent if the blockquote
is not editable.

* LayoutTests/editing/execCommand/indent-user-select-all-blockquotes-expected.txt: Added.
* LayoutTests/editing/execCommand/indent-user-select-all-blockquotes.html: Added.
* Source/WebCore/editing/IndentOutdentCommand.cpp:
(WebCore::IndentOutdentCommand::indentIntoBlockquote):

Canonical link: https://commits.webkit.org/256749@main