Address FIXMEs in process-css-properties.py https://bugs.webkit.org/show_bug.cgi?id=247861 rdar://102292266
Reviewed by Darin Adler.

Address the FIXMEs I left in process-css-properties.py to make the
transition from the old script easier. This mostly was removal of
unnecessary extra sorting that was done to maintain character for
character parity with the old script.

Additionally, this removes a bunch of unnecessary list allocations
by converting list-comprehension expressions (i.e. [x for x in foo])
into generator-comprehension expressions (i.e. (x for x in foo)) where
materializing the list early was not necessary or useful. This allows
getting rid of a bunch of caching as well, which is nice.

* Source/WebCore/css/process-css-properties.py:

Canonical link: https://commits.webkit.org/256763@main