[iOS] PiP-then-Fullscreen results in broken controls in fullscreen mode https://bugs.webkit.org/show_bug.cgi?id=247991 rdar://100442052
Reviewed by Eric Carlson.

WebFullScreenManager's m_element ivar is never cleared, which means the next time a video plays
within a page, the WebFullScreenManager thinks its still in fullscreen, and the playing element
is marked as the PIPStandbyElement. This causes the element's own m_videoFullscreenStandby ivar
to be set to true, and when that element is taken into fullscreen, it's fullscreen controls are
explicitly hidden.

No test was included because TestWebKitAPI is not allowed to enter fullscreen or PiP as its not
a "real" application.

* Source/WebKit/WebProcess/FullScreen/WebFullScreenManager.cpp:
(WebKit::WebFullScreenManager::supportsFullScreen):
(WebKit::eventsToObserve):
(WebKit::WebFullScreenManager::setElement):
(WebKit::WebFullScreenManager::clearElement):

Canonical link: https://commits.webkit.org/256812@main