[Web Animations] Make WPT test at animation-model/keyframe-effects/effect-value-context.html pass reliably https://bugs.webkit.org/show_bug.cgi?id=186490 rdar://41000137
Reviewed by Antti Koivisto.

We added support for recomputing keyframes when the computed font-size changes in bug 237357, allowing
any keyframe values depending on font-size to produce the expected values. However, we have a bug when
font-size changes on a parent element where the logic under ComputedStyleExtractor::propertyValue() will
not invalidate style because the code does not think the style needs updating.

So we update hasValidStyleForProperty() to account for animations affecting the property for which
the computed style is requested.

* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-filling-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/boxShadow-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/shapeOutside-expected.txt:
* Source/WebCore/animation/KeyframeEffectStack.cpp:
(WebCore::KeyframeEffectStack::containsProperty const):
* Source/WebCore/animation/KeyframeEffectStack.h:
* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::hasValidStyleForProperty):

Canonical link: https://commits.webkit.org/256889@main