[WPE] built-product-archive not longer includes cog in the release.zip archive https://bugs.webkit.org/show_bug.cgi?id=248233
Reviewed by Michael Catanzaro.

Since cog moved to Meson build system, the paths it uses for storing the built
products on the disk have changed. Instead of using a predefined names for
the paths just add to the product-archive the files (binaries) 'cog', 'cogctl'
as well as any library.

* Tools/CISupport/built-product-archive:

Canonical link: https://commits.webkit.org/256941@main