[web-animations] word-spacing should support animating between percentage and fixed values https://bugs.webkit.org/show_bug.cgi?id=248218
Reviewed by Tim Nguyen.

While we support percentage values for word-spacing, we do not support blending between fixed and percentage values,
so we add the <length-percentage> flag to its animation wrapper. And now that we may produce calculated values as a
result, we account for non-fixed values when serializing the computed style for that property.

There remains the issue where we fail to parse calculated values for word-spacing and this causes a few remaining
FAIL results, this will be dealt with in bug 248220.

* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/animation-types/accumulation-per-property-002-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/animation-types/addition-per-property-002-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/animation-types/interpolation-per-property-002-expected.txt:
* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):
* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::ComputedStyleExtractor::valueForPropertyInStyle):

Canonical link: https://commits.webkit.org/256951@main