[web-animations] correctly accumulate and clamp filter values when blending with "none" https://bugs.webkit.org/show_bug.cgi?id=248272
Reviewed by Simon Fraser.

Blending a valid operation with a "none" operation is done using a passthrough value in the
WebCore filter blending code. However, when we added support for accumulation and clamping in
bug 248235, we only did so for the case when blending between two valid operations. We now use
the blendAmounts() method in the passthrough cases as well.

* LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/animation/filter-interpolation-003-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-iteration-composite-operation-expected.txt:
* Source/WebCore/platform/graphics/filters/FilterOperation.cpp:
(WebCore::BasicColorMatrixFilterOperation::blend):
(WebCore::BasicComponentTransferFilterOperation::blend):

Canonical link: https://commits.webkit.org/256976@main